Add pyfunc utils module
