+++ mlflow/pyfunc/__init__.py
@@ -85,8 +85,11 @@ import sys
 from mlflow.tracking.fluent import active_run, log_artifacts
 from mlflow import tracking
 from mlflow.models import Model
-from mlflow.utils import PYTHON_VERSION, get_major_minor_py_version
+from mlflow.pyfunc.utils import _get_code_dirs,\
+        _warn_potentially_incompatible_py_version_if_necessary
+from mlflow.utils import PYTHON_VERSION 
 from mlflow.utils.file_utils import TempDir, _copy_file_or_tree
+from mlflow.utils.model_utils import _get_flavor_configuration
 from mlflow.exceptions import MlflowException
 from mlflow.protos.databricks_pb2 import INVALID_PARAMETER_VALUE
 
@@ -130,26 +133,6 @@ def add_to_model(model, loader_module, data=None, code=None, env=None):
     return model.add_flavor(FLAVOR_NAME, **parms)
 
 
-def _load_model_conf(path, run_id=None):
-    """Load a model configuration stored in Python function format."""
-    if run_id:
-        path = tracking.utils._get_model_log_dir(path, run_id)
-    conf_path = os.path.join(path, "MLmodel")
-    model = Model.load(conf_path)
-    if FLAVOR_NAME not in model.flavors:
-        raise Exception("Format '{format}' not found not in {path}.".format(format=FLAVOR_NAME,
-                                                                            path=conf_path))
-    return model.flavors[FLAVOR_NAME]
-
-
-def _load_model_env(path, run_id=None):
-    """
-        Get ENV file string from a model configuration stored in Python Function format.
-        Returned value is a model-relative path to a Conda Environment file,
-        or None if none was specified at model save time
-    """
-    return _load_model_conf(path, run_id).get(ENV, None)
-
 
 def load_pyfunc(path, run_id=None, suppress_warnings=False):
     """
@@ -161,9 +144,9 @@ def load_pyfunc(path, run_id=None, suppress_warnings=False):
                               loading process will be suppressed. If False, these warning messages
                               will be emitted.
     """
-    if run_id:
+    if run_id is not None:
         path = tracking.utils._get_model_log_dir(path, run_id)
-    conf = _load_model_conf(path)
+    conf = _get_flavor_configuration(model_path=path, flavor_name=FLAVOR_NAME)
     model_py_version = conf.get(PY_VERSION)
     if not suppress_warnings:
         _warn_potentially_incompatible_py_version_if_necessary(model_py_version=model_py_version)
@@ -172,28 +155,6 @@ def load_pyfunc(path, run_id=None, suppress_warnings=False):
         sys.path = [code_path] + _get_code_dirs(code_path) + sys.path
     data_path = os.path.join(path, conf[DATA]) if (DATA in conf) else path
     return importlib.import_module(conf[MAIN])._load_pyfunc(data_path)
-
-
-def _warn_potentially_incompatible_py_version_if_necessary(model_py_version):
-    if model_py_version is None:
-        _logger.warning(
-            "The specified model does not have a specified Python version. It may be"
-            " incompatible with the version of Python that is currently running: Python %s",
-            PYTHON_VERSION)
-    elif get_major_minor_py_version(model_py_version) != get_major_minor_py_version(PYTHON_VERSION):
-        _logger.warning(
-            "The version of Python that the model was saved in, `Python %s`, differs"
-            " from the version of Python that is currently running, `Python %s`,"
-            " and may be incompatible",
-            model_py_version, PYTHON_VERSION)
-
-
-def _get_code_dirs(src_code_path, dst_code_path=None):
-    if not dst_code_path:
-        dst_code_path = src_code_path
-    return [(os.path.join(dst_code_path, x))
-            for x in os.listdir(src_code_path) if not x.endswith(".py") and not
-            x.endswith(".pyc") and not x == "__pycache__"]
 
 
 def spark_udf(spark, path, run_id=None, result_type="double"):
